feat(swap): add Market/Limit tab row above the swap form
Introduce the Advanced Swap mode tabs (phase 0 of #4099). The swap
screen now renders a Market/Limit tab row beneath the header, with each
tab holding its own content:

- Market: the existing swap form (moved to MarketSwapForm)
- Limit: placeholder content (limit execution is out of scope here)
